requirements.txt
asyncore-wsgi==0.0.9
autopep8==1.4.4
backcall==0.1.0
bottle==0.12.17
certifi==2019.6.16
chardet==3.0.4
decorator==4.4.0
defusedxml==0.6.0
Django==2.2.4
django-allauth==0.40.0
idna==2.8
ipython==7.7.0
ipython-genutils==0.2.0
jedi==0.15.1
oauthlib==3.1.0
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
pkg-resources==0.0.0
prompt-toolkit==2.0.9
ptyprocess==0.6.0
pycodestyle==2.5.0
Pygments==2.4.2
PyJWT==1.7.1
python-dotenv==0.10.3
python3-openid==3.1.0
pytz==2019.2
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
sqlparse==0.3.0
traitlets==4.3.2
urllib3==1.25.3
wcwidth==0.1.7
web-pdb==1.5.1
